Created delete nurses backend and basic frontend capability.

+++ backend/controllers/deleteController.js
@@ -1,16 +1,22 @@
 const Nurse = require('../database/models/nurse');
+const mongoose = require('mongoose');
 
-const deleteNurse = async (id) => {
+const deleteNurse = async (req, res) => {
     try {
-        const res = await Nurse.findByIdAndDelete(id);
+        const { id } = req.params;
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({success: false, message: "Invalid nurse ID."})
+        }
+        const result = await Nurse.findByIdAndDelete(id);
 
-        if (res) {
+        if (result) {
             res.status(200).json({success: true, message: `Object with id ${id} successfully deleted.` })
         } else {
             res.status(404).json({success: false, message: `Object with id ${id} not found`})
         }
     } catch (err) {
             console.error('Error deleting object', err)
+            res.status(500).json({success: false, message: "internal server error"})
     }
 }
